Reorganize column display options types
Prepare these types to be used by tables and explorations
