+++ mathesar_ui/src/stores/abstract-types/type-configs/datetime.ts
@@ -1,9 +1,8 @@
-import {
-  type Column,
-  type DateFormat,
-  type TimeFormat,
-  getColumnMetadataValue,
-} from '@mathesar/api/rpc/columns';
+import type {
+  DateFormat,
+  TimeFormat,
+} from '@mathesar/api/rpc/_common/columnDisplayOptions';
+import { type Column, getColumnMetadataValue } from '@mathesar/api/rpc/columns';
 import { iconUiTypeDateTime } from '@mathesar/icons';
 import type { FormValues } from '@mathesar-component-library/types';
