+++ mathesar_ui/src/stores/table-data/columns.ts
@@ -2,10 +2,10 @@ import { type Readable, derived, writable } from 'svelte/store';
 
 import type { RequestStatus } from '@mathesar/api/rest/utils/requestUtils';
 import { api } from '@mathesar/api/rpc';
+import type { ColumnMetadata } from '@mathesar/api/rpc/_common/columnDisplayOptions';
 import type {
   Column,
   ColumnCreationSpec,
-  ColumnMetadata,
   ColumnPatchSpec,
 } from '@mathesar/api/rpc/columns';
 import type { Database } from '@mathesar/models/Database';
